feat: Added Basic Tilemap anShooting functionality locally
