+++ Assets/New Tile Palette.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ee3adb3cd5a85cd4d8273b7d11db199e
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
