+++ Assets/Particles.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2d257b5c38f499844a9e2a18607523ab
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
