+++ Assets/Particles/Particle System.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6543df18a3460644ab1d9953d7b478f2
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
