+++ Assets/Prefabs/Bullet.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a8a33bac94364bd45803e7a68dfe4e4b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
