+++ Assets/Scripts/Bullet.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 06c1857dde2fb184db5a7522c211b1b1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
