+++ Assets/Scripts/Player/PlayerShooting.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: ccaff4d78e46ad840b4c61df9091a448
