+++ Assets/Square.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1cdbc691ae0decd47a10ae0f50bee5c8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
